Refactor Boxed class
